Finish porting code to the Teensy
