+++ teensy/.gitignore
@@ -1,4 +1,3 @@
 .pio
 CMakeListsPrivate.txt
 cmake-build-*/ 
-.idea
