+++ teensy/src/motor.cpp
